Fix feature extraction
